requirements.txt
# Core dependencies
google-generativeai>=0.3.2
python-dotenv>=1.0.0
pyyaml>=6.0.1
asyncio-throttle>=1.0.2

# Data handling and analysis
pandas>=2.0.0
numpy>=1.24.0
json5>=0.9.0

# Visualization and metrics
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Utilities
tqdm>=4.65.0
colorama>=0.4.6
rich>=13.4.0

# Optional: For local development
jupyter>=1.0.0
ipykernel>=6.24.0 